Migrate form handlers into formevents.js
